requirements.txt
# === FastAPI & Server ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# === Database (PostgreSQL) ===
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0

# === Pydantic & Settings ===
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# === Authentication & Security ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17

# === HTTP Client (for external APIs) ===
httpx>=0.28.0

# === Environment Variables ===
python-dotenv>=1.0.1

# === Logging & Utilities ===
structlog>=24.4.0

# === Testing ===
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.28.0

# === Development Tools ===
ruff>=0.8.0
mypy>=1.13.0